- Rename files and variables for clarity - Remove unused set functions
